変更 - ファイルの場所を移設し、ClientComponent化させ、AppRouter対応
